Maine personal exemption Fixes #1297
