--- policyengine_us/parameters/gov/states/me/tax/income/credits/personal_exemption/income_limit/head_of_household.yaml
@@ -1,10 +1,0 @@
-description: Maine income limit for head of household filing status for personal exemption credit.
-metadata:
-  type: single_amount
-  unit: /1
-  label: Maine Personal Exemption Income Limit (Head of Household)
-  reference:
-    - title: Maine Form 1040ME Instructions
-      href: https://www.maine.gov/revenue/sites/maine.gov.revenue/files/inline-files/21_1040me_book_gen_instr_revMay2022.pdf
-values:
-  2022-01-01: 307_450
