--- policyengine_us/parameters/gov/states/me/tax/income/credits/personal_exemption/income_limit/separate.yaml
@@ -1,10 +1,0 @@
-description: Maine income limit for separate filing status for personal exemption credit.
-metadata:
-  type: single_amount
-  unit: /1
-  label: Maine Personal Exemption Income Limit (Separate)
-  reference:
-    - title: Maine Form 1040ME Instructions
-      href: https://www.maine.gov/revenue/sites/maine.gov.revenue/files/inline-files/21_1040me_book_gen_instr_revMay2022.pdf
-values:
-  2022-01-01: 167_700
