--- policyengine_us/tests/policy/baseline/gov/states/me/tax/income/credits/me_personal_exemption.yaml
@@ -1,29 +1,0 @@
-- name: Outside Maine, no personal exemption.
-  period: 2022
-  absolute_error_margin: 0
-  input:
-    state_code: NY
-    filing_status: SINGLE
-  output:
-    me_personal_exemption: 0
-
-- name: If $0 income, personal exemption is x
-  period: 2022
-  absolute_error_margin: 0
-  input:
-    state_code: ME
-    me_agi: 0
-    filing_status: SINGLE
-    me_personal_exemptions: 1
-  output:
-    me_personal_exemption: 4_300
-
-- name: High income filers don't get a personal exemption.
-  period: 2022
-  absolute_error_margin: 0
-  input:
-    state_code: ME
-    filing_status: SINGLE
-    me_agi: 297_500
-  output:
-    me_personal_exemption: 0
