+++ policyengine_us/tests/policy/baseline/gov/states/me/tax/income/deductions/me_personal_exemption.yaml
@@ -1,0 +1,53 @@
+- name: Outside Maine, no personal exemption.
+  period: 2021
+  absolute_error_margin: 0
+  input:
+    state_code: NY
+    filing_status: SINGLE
+  output:
+    me_personal_exemption_deduction: 0
+
+- name: If $0 income, personal exemption is $4,300.
+  period: 2021
+  absolute_error_margin: 0
+  input:
+    state_code: ME
+    me_agi: 0
+    filing_status: SINGLE
+    me_personal_exemptions: 1
+  output:
+    me_personal_exemption_deduction: 4_300
+
+- name: If you have an income below the phaseout start, you get the full personal exemption.
+  period: 2021
+  absolute_error_margin: 0
+  input:
+    state_code: ME
+    filing_status: SINGLE
+    me_agi: 297_500
+    me_personal_exemptions: 1
+  output:
+    me_personal_exemption_deduction: 4_300
+
+- name: If you have an income between the phaseout start and end, you get a partial credit.
+  period: 2021
+  absolute_error_margin: 0
+  input:
+    state_code: ME
+    filing_status: SINGLE
+    me_agi: 400_000
+    me_personal_exemptions: 1
+  output:
+    # Excess = $120,500 / $125,000 = 0.964
+    me_personal_exemption_deduction: 154.8 # $4,300 - $4,145.2
+
+- name: High income filers get no personal exemption.
+  period: 2021
+  absolute_error_margin: 0
+  input:
+    state_code: ME
+    filing_status: SINGLE
+    me_agi: 404_500 # $279,500 + $125,000 = $404,500
+    me_personal_exemptions: 1
+  output:
+    me_personal_exemption_deduction: 0
