+++ policyengine_us/variables/gov/states/me/adjusted_gross_income/me_agi.py
@@ -1,0 +1,10 @@
+from policyengine_us.model_api import *
+
+
+class me_agi(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Maine adjusted gross income"
+    unit = USD
+    definition_period = YEAR
+    defined_for = StateCode.ME
