+++ policyengine_us/variables/gov/states/me/tax/income/deductions/personal_exemption/me_personal_exemptions.py
@@ -1,0 +1,10 @@
+from policyengine_us.model_api import *
+
+
+class me_personal_exemptions(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Maine personal exemptions"
+    unit = USD
+    definition_period = YEAR
+    defined_for = StateCode.ME
